fix(contract): use bigint instead of int for GenLayer storage types, deploy to studionet

+++ contract/fact_checker.py
@@ -103,17 +103,17 @@ class FactCheckRecord:
     claim: str                # raw claim text (max 500 chars)
     source_url: str           # primary URL submitted by user
     verdict: str              # TRUE | FALSE | MISLEADING | UNVERIFIABLE
-    confidence: int           # 0-100
+    confidence: bigint        # 0-100
     explanation: str          # LLM-generated 2-3 sentence reasoning
     sources_checked: DynArray[str]
-    timestamp: int            # block timestamp
+    timestamp: bigint         # block timestamp
     submitter: str            # wallet address
 
 
 class FactChecker(gl.Contract):
     checks: TreeMap[str, FactCheckRecord]
-    total_checks: int
-    verdicts_by_type: TreeMap[str, int]
+    total_checks: bigint
+    verdicts_by_type: TreeMap[str, bigint]
 
     def __init__(self):
         self.total_checks = 0
